refactor(mastery): split monster-system into mastery/ submodules with shim (U7)
