+++ src/platform/game/mastery/punctuation.js
@@ -1,0 +1,207 @@
+import { MONSTERS } from '../monsters.js';
+import { PUNCTUATION_CURRENT_RELEASE_ID } from '../../../subjects/punctuation/service-contract.js';
+import {
+  branchForMonster,
+  DEFAULT_SYSTEM_ID,
+  ensureMonsterBranches,
+  isPlainObject,
+  masteredList,
+  PUNCTUATION_GRAND_MONSTER_ID,
+  PUNCTUATION_MONSTER_IDS,
+  releaseIdForEntry,
+  saveMonsterState,
+  toastBodyFor,
+} from './shared.js';
+
+function punctuationMasteredList(entry, releaseId = null) {
+  const mastered = masteredList(entry);
+  const scopedReleaseId = releaseIdForEntry(entry, releaseId);
+  if (!scopedReleaseId) return mastered;
+  const releasePrefix = `punctuation:${scopedReleaseId}:`;
+  return mastered.filter((key) => key.startsWith(releasePrefix));
+}
+
+function punctuationMasteredCount(entry, releaseId = null) {
+  const mastered = masteredList(entry);
+  if (mastered.length) return punctuationMasteredList(entry, releaseId).length;
+  const count = Number(entry?.masteredCount);
+  return Number.isFinite(count) && count > 0 ? Math.floor(count) : 0;
+}
+
+function punctuationTotal(entry, fallback = 1) {
+  const count = Number(entry?.publishedTotal);
+  return Number.isFinite(count) && count > 0 ? Math.floor(count) : Math.max(1, Number(fallback) || 1);
+}
+
+function punctuationStageFor(mastered, total) {
+  const denominator = Math.max(1, Number(total) || 1);
+  const ratio = Math.max(0, Math.min(1, (Number(mastered) || 0) / denominator));
+  if (ratio >= 1) return 4;
+  if (ratio >= 2 / 3) return 3;
+  if (ratio >= 1 / 3) return 2;
+  if (ratio > 0) return 1;
+  return 0;
+}
+
+export function activePunctuationMonsterSummaryFromState(state = {}) {
+  return punctuationMonsterSummaryFromState(state)
+    .filter((entry) => entry.progress.caught || entry.progress.mastered > 0);
+}
+
+export function progressForPunctuationMonster(state, monsterId, { publishedTotal = null, releaseId = PUNCTUATION_CURRENT_RELEASE_ID } = {}) {
+  const entry = isPlainObject(state?.[monsterId]) ? state[monsterId] : { mastered: [], caught: false };
+  const mastered = punctuationMasteredCount(entry, releaseId);
+  const total = punctuationTotal(entry, publishedTotal || MONSTERS[monsterId]?.masteredMax || 1);
+  return {
+    mastered,
+    publishedTotal: total,
+    stage: punctuationStageFor(mastered, total),
+    level: Math.min(10, Math.round((mastered / Math.max(1, total)) * 10)),
+    caught: mastered >= 1,
+    branch: branchForMonster(state, monsterId),
+    masteredList: punctuationMasteredList(entry, releaseId),
+  };
+}
+
+function buildPunctuationEvent({
+  learnerId,
+  kind,
+  monsterId,
+  previous,
+  next,
+  releaseId,
+  clusterId,
+  rewardUnitId,
+  masteryKey,
+  createdAt = Date.now(),
+} = {}) {
+  const monster = MONSTERS[monsterId];
+  return {
+    id: `reward.monster:${learnerId || 'default'}:punctuation:${releaseId}:${clusterId}:${rewardUnitId}:${monsterId}:${kind}`,
+    type: 'reward.monster',
+    kind,
+    learnerId,
+    subjectId: 'punctuation',
+    systemId: DEFAULT_SYSTEM_ID,
+    releaseId,
+    clusterId,
+    rewardUnitId,
+    masteryKey,
+    monsterId,
+    monster,
+    previous,
+    next,
+    createdAt,
+    toast: {
+      title: monster?.name || 'Reward update',
+      body: toastBodyFor(kind),
+    },
+  };
+}
+
+function punctuationEventFromTransition(payload, previous, next) {
+  if (!previous.caught && next.caught) {
+    return buildPunctuationEvent({ ...payload, kind: 'caught', previous, next });
+  }
+  if (next.stage > previous.stage) {
+    return buildPunctuationEvent({ ...payload, kind: next.stage === 4 ? 'mega' : 'evolve', previous, next });
+  }
+  if (next.level > previous.level) {
+    return buildPunctuationEvent({ ...payload, kind: 'levelup', previous, next });
+  }
+  return null;
+}
+
+export function recordPunctuationRewardUnitMastery({
+  learnerId,
+  releaseId,
+  clusterId,
+  rewardUnitId,
+  masteryKey,
+  monsterId,
+  publishedTotal = 1,
+  aggregateMonsterId = PUNCTUATION_GRAND_MONSTER_ID,
+  aggregatePublishedTotal = 1,
+  createdAt = Date.now(),
+  gameStateRepository,
+  random = Math.random,
+} = {}) {
+  if (!MONSTERS[monsterId] || !masteryKey) return [];
+  const scopedReleaseId = typeof releaseId === 'string' && releaseId ? releaseId : PUNCTUATION_CURRENT_RELEASE_ID;
+  const expectedMasteryKey = `punctuation:${scopedReleaseId}:${clusterId}:${rewardUnitId}`;
+  if (masteryKey !== expectedMasteryKey) return [];
+  const before = ensureMonsterBranches(learnerId, gameStateRepository, {
+    random,
+    monsterIds: PUNCTUATION_MONSTER_IDS,
+  });
+  const directEntry = isPlainObject(before[monsterId]) ? before[monsterId] : { mastered: [], caught: false };
+  const directMastered = masteredList(directEntry);
+  if (directMastered.includes(masteryKey)) return [];
+
+  const aggregateEntry = isPlainObject(before[aggregateMonsterId]) ? before[aggregateMonsterId] : { mastered: [], caught: false };
+  const aggregateMastered = masteredList(aggregateEntry);
+  const beforeDirect = progressForPunctuationMonster(before, monsterId, { publishedTotal, releaseId: scopedReleaseId });
+  const beforeAggregate = progressForPunctuationMonster(before, aggregateMonsterId, { publishedTotal: aggregatePublishedTotal, releaseId: scopedReleaseId });
+
+  const after = {
+    ...before,
+    [monsterId]: {
+      ...directEntry,
+      caught: true,
+      releaseId: scopedReleaseId,
+      publishedTotal,
+      mastered: [...directMastered, masteryKey],
+    },
+    [aggregateMonsterId]: {
+      ...aggregateEntry,
+      caught: true,
+      releaseId: scopedReleaseId,
+      publishedTotal: aggregatePublishedTotal,
+      mastered: aggregateMastered.includes(masteryKey)
+        ? aggregateMastered
+        : [...aggregateMastered, masteryKey],
+    },
+  };
+
+  const afterDirect = progressForPunctuationMonster(after, monsterId, { publishedTotal, releaseId: scopedReleaseId });
+  const afterAggregate = progressForPunctuationMonster(after, aggregateMonsterId, { publishedTotal: aggregatePublishedTotal, releaseId: scopedReleaseId });
+  saveMonsterState(learnerId, after, gameStateRepository);
+
+  const events = [];
+  const directEvent = punctuationEventFromTransition({
+    learnerId,
+    monsterId,
+    releaseId: scopedReleaseId,
+    clusterId,
+    rewardUnitId,
+    masteryKey,
+    createdAt,
+  }, beforeDirect, afterDirect);
+  if (directEvent) events.push(directEvent);
+
+  const aggregateEvent = punctuationEventFromTransition({
+    learnerId,
+    monsterId: aggregateMonsterId,
+    releaseId: scopedReleaseId,
+    clusterId: 'published_release',
+    rewardUnitId,
+    masteryKey,
+    createdAt,
+  }, beforeAggregate, afterAggregate);
+  if (aggregateEvent) events.push(aggregateEvent);
+
+  return events;
+}
+
+export function punctuationMonsterSummaryFromState(state = {}, { clusterTotals = {}, aggregateTotal = 1, releaseId = PUNCTUATION_CURRENT_RELEASE_ID } = {}) {
+  return PUNCTUATION_MONSTER_IDS.map((monsterId) => ({
+    subjectId: 'punctuation',
+    monster: MONSTERS[monsterId],
+    progress: progressForPunctuationMonster(state, monsterId, {
+      publishedTotal: monsterId === PUNCTUATION_GRAND_MONSTER_ID
+        ? aggregateTotal
+        : (clusterTotals[monsterId] || MONSTERS[monsterId]?.masteredMax || 1),
+      releaseId,
+    }),
+  }));
+}
